+++ alarm/exceptions.py
@@ -1,0 +1,2 @@
+class AlarmFileException(Exception):
+    """To be used in case of an invalid alarm file"""
